add: mock the /api calls
