+++ apps/golden-sample-app/proxy.mocks.conf.json
@@ -1,0 +1,5 @@
+{
+  "context": "/api",
+  "target": "http://localhost:9999",
+  "secure": false
+}
