+++ apps/golden-sample-app/src/environments/environment.mocks.ts
@@ -1,0 +1,35 @@
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { Provider } from '@angular/core';
+import { AuthConfig } from 'angular-oauth2-oidc';
+import { AchPositivePayInterceptor } from '../app/interceptors/ach-positive-pay.interceptor';
+import { Environment } from './type';
+
+const mockProviders: Provider[] = [
+  {
+    provide: HTTP_INTERCEPTORS,
+    useClass: AchPositivePayInterceptor,
+    multi: true,
+  },
+];
+
+export const environment: Environment = {
+  production: false,
+  apiRoot: '/api',
+  mockProviders,
+  locales: ['en-US', 'nl-NL'],
+  common: {
+    designSlimMode: false,
+  },
+};
+
+export const authConfig: AuthConfig = {
+  issuer:
+    'https://identity.prd.sdbxaz.azure.backbaseservices.com/auth/realms/customer',
+  redirectUri: document.baseURI,
+  clientId: 'bb-web-client',
+  responseType: 'code',
+  scope: 'openid',
+  requireHttps: false,
+  showDebugInformation: true,
+  logoutUrl: document.baseURI + 'logout',
+};
